Add flag to specify prebuilt OMPT is TR6.
The only way to specify TR6 is to use -ompt=download-tr6. Using
-ompt=/path/to/ompt assumes TR4, and there is no version macro in the
package for easily detecting the version. This adds a flags, -ompt-tr6,
which forces the processing of a path as TR6 rather than TR4. This is
used in TAU Commander, which downloads its own copy of OMPT rather than
using TAU's download flags.


Former-commit-id: d96093c95bff2cfac65f2365264f20a5b31794df