Only set time for context events on "fake" threads
For events that happen on actual (non-device) threads, stage_gpu_event
sets the GPU vim sets the GPU timestamp, but, because the thread is
non-fake, that timestamp is never actually used. So when recording CUDA
memcpy events on those threads, we have to use the current time for the
context event, not the GPU timestamp, since that is ignored.


Former-commit-id: 712bdcf7027053ef84361425d9dd62aef6e0f5cd