OMPT Plugins can now use callbacks unused by TAU
OMPT plugins could not use an OMPT callback that is not used by TAU. Now
if a callback is not used by TAU it will still be register for the
plugins, if at least one plugin requested it. TAU will ignore it's part
of the callback, and only invoke the plugins' callbacks.

Note: This commit adds a one level of indentation in a lot of routines
resulting in a lot of line changed simply because their indentation
changed.

Former-commit-id: 5f349e161341220dac0879337ee332cf9774adba