Very valuable lesson about reserving memory.
This is a performance fix - the ADIOS2 trace plugin was
reserving memory during initialization to hold the trace buffer.
The *assumption* was that this would help performance.
Ironically, it made performance worse on Summit - reserving
a block of memory on thread 0 that gets written by other threads
and then read by another thread was poison for performance.
This was discovered by testing with jemalloc and finding that
the performance problems went away when a thread-friendly
heap manager was used.  selectively commenting out code led
to the discovery.

Beware premature optimization!!


Former-commit-id: 214b10304c51eb74ca69a31da640eaf61c904993