Because of unintended consequences, the TAU_RECYCLE_THREADS default is now 0.
Unfortunately, thread IDs are reused for pthreads and for cuda
streams.  They end up getting mixed together - thread IDs that are
used for pthreads get reused for CUDA streams.  So, until this can be
fixed in a more robust way, return to the default thread handling
in TAU for now.


Former-commit-id: ad50e69505590ec2327e98a044fdb53f7b0cb3ba