Because of unintended consequences, the TAU_RECYCLE_THREADS default is now 0.
Unfortunately, thread IDs are reused for pthreads and for cuda
streams.  They end up getting mixed together - thread IDs that are
used for pthreads get reused for CUDA streams.  So, until this can be
fixed in a more robust way, return to the default thread handling
in TAU for now.


Former-commit-id: 17c51e94f9062b1fb79c7adcf375a1e087b6f9a5