Fix hang when sampling interrupts syscalls
When a signal is received, a syscall is interrupted.
This commit makes two changes to ensure that a syscall
interrupted by TAU is eventually able to complete:

1) Sets SA_RESTART in sigaction unless the application has previously
registered a sigaction with that flag explictly unset. For resumable
syscalls (e.g., I/O operations), the call will automatically resume
when TAU's signal handler returns.

2) When a thread calls fork(), which is non-resumable, disable the
sampling timer on that thread specifically by disarming the itimer.
Upon returning from fork() on the parent, re-arm the timer.

Together, these fixes prevent the problem where a hang occurs during
TensorFlow import when Python and EBS are used together.


Former-commit-id: d2c51b6a2912829690d265bf2e0cc605fd4a0dfd