Workaround for macOS codesign bug
On arm64, macOS *requires* that all executables be signed, which
the compiler will do automatically. However, due to an OS bug,
if an existing, already-signed executable is replaced, the old
signature will still be in an OS-level cache and the executable
will fail to run.

This works around the issue by copying the arm64_apple directory
to a temporary directory, removing the original, and renaming
the temporary to replace the original. Changing the inode of the files
invalidates the cache.


Former-commit-id: 2dccc61996f13e562bb750b19798af898f26279b