Fix timer overlap when CUPTI host timer throttled
The code in TauGpu.cpp was, on exit from a GPU host timer,
simply stopping whatever was on the timer stack for the relevant
thread. In the event that a timer had been throttled, it would not
have been started and therefore this would stop the wrong timer,
leading to a subsequent overlap when the stop for the timer
that had been on the stack was called.

This appears to fix an issue where TensorFlow applications would give
an overlapping timer error unless TAU_THROTTLE was set to 0.


Former-commit-id: 51262f09735f67c943bcb6072e7fce0a92416139