Fix race condition in sampling init
During sampling init, TAU_EBS_SOURCE is checked to see if it is TIME or
itimer. If not, sampling is not started. Later during init,
TAU_EBS_SOURCE is changed if something other than TIME is the first
metric. The first thread to make it to that point then changes
TAU_EBS_SOURCE and prevents other threads from starting sampling. This
commit fixes that problem by only checking against the original value of
TAU_EBS_SOURCE, not what it is changed to while TAU is running.

Also corrects an issue where TAU_EBS_SOURCE was not written as metadata
to the profile.


Former-commit-id: 9aaac0a939c1fffad8fa9a7a324b2873c75d319a