Fix crash when using two library wrappers together
Libraries generated by tau_wrap were storing the original library name
in 'const char * tau_orig_libname'. Because this was not declared
static, there was a single tau_orig_libname for the entire process. If
multiple wrappers generated by tau_wrap were used simultaneously, the
later-loading wrapper would overwrite the name of the previous one,
breaking the wrappers for all but the last-loaded wrapper.

This is fixed by declaring the variable static.


Former-commit-id: f7d63e0c701f92ce98d2d548286e7ef769552c64