Do not use the full path in filenames in the LLVM plugin, for compatibility purpose.
Previous versions of the plugin were using only the names of the files to include/exclude, such as foo.c.
With the modifications from the latest commit, we can use the full absolute name, such as /home/toto/foo.c.
However, this might break things, because it requires using the full absolute names in the select files.
Hence, for compatibility purposes, I commented out the lines that get the absolute path, and we are only
using the names of the files.


Former-commit-id: 69a25f7b3f4781d8c1e046a49d06ee81ad40282c