Query Python interpreter for Python paths
When building with -python, TAU's configure script would try to find an
include directory and libpython relative to the interpreter. It would
pick the first one listed by ls. This doesn't always work correctly,
because multiple Pythons can be installed into the same directory, in
which case TAU could pick a different Python include and library version
than the interpreter.

This changes `configure` to execute python and use the module
`sysconfig` to get the Python include directory, library directory, and
library name. If that fails, it falls back to the old approach.

Also, -python and -python3 now take optional arguments to specify a
Python interpreter by name, if a version other than the default is
needed.


Former-commit-id: b952846b32167c5012734ccf84fe6e5e03b84805