Allow another thread to start sampling on thread 0.
When MPI is used, sampling initialization is deferred until MPI_Init is
called. The code that handled this assumed that MPI_Init would be called
from thread 0, but the MPI spec does not require this, and Horovod in
fact calls MPI_Init_thread from thread 1. In that case, sampling on
thread 0 would never start. Resolve this by adding thread 0 to the
deferred threads list when TAU is initialized.

Additionaly, there was a problem with one thread starting sampling for
another thread: because the tau_sampling_flags was thread-local, if
thread 1 started sampling on thread 0, thread 0's flags would not be
initialized first. This is fixed by changing tau_sampling_flags from
being a thread-local only to being a thread-local cache onto a map.


Former-commit-id: 4cbbb6c2bf30dfb5874ce297394a26c73df68236