Fix version detection in TauBfd
TauBfd was using TAU_BFD to condition features on the version of BFD,
but TAU_BFD was just always set to 1 if TAU was built with BFD, so
version-conditional feature were always disabled. In particular,
discriminator support was always disabled and the older API for line
resolution was always being used, regardless of the BFD version.
Instead, check whether macros are defined before using them.


Former-commit-id: 998f649acdbbaea6fba820ef4e68d4a1a15be7a7