Use PBS_NODENUM to get rank, not PBS_TASKNUM
In the absence of MPI, TAU will check several environment variables to
try to get a rank from the scheduler. One of these was PBS_TASKNUM, but
that was the wrong PBS variable to be checking -- it gives the task
within a job, not the node. The node is in PBS_NODENUM. We now use that
instead.


Former-commit-id: f1e594ad9291be8d7ce27249efe243d79e60117a