Don't try to get rank from PBS vars
None of the PBS env vars contain a correct rank on Aurora, but are set
inside interactive sessions, yielding a rank of 1 for the single rank
when running e.g. tau_python on a compute node in an interactive session
without mpirun. Remove PBS_TASKNUM. Add PALS_RANKID, which was checked in
ZeroSum but not TAU.


Former-commit-id: da582d0389081cea09c97e7613d6d66e9360279e