Make sure top-level timer gets started on Intel GPU task.
If TAU is run with -l0, and the Level Zero API is called, but no
kernels are run, the profile for the Intel GPU device thread
will not be written, while the higher-numbered host thread will.
This fixes the issue by starting the top-level timer on the
device thread as soon as any callback is received, whether a
host or device callback.


Former-commit-id: e5807430000e9055be3b05c4eebefb7c380bc6e8