Ignore threads started before shmem_init()
When TAU is built with -shmem, ignore threads that are launched before
shmem_init() returns. HPC-X loads a shared library during shmem_init()
which has a constructor function which spawns a thread. Loading a shared
library acquires a lock which is also required for initializing
thread-local storage, so if TAU wraps this thread, the program will
deadlock as TAU tries to acquire a lock that dlopen() holds and won't
release until the constructor function returns.


Former-commit-id: 5540765a68e5c406e8bb9f15ba8db1f54804899f