Resolve samples with ranges from ITTNotify JIT Events
In addition to synchronous start/stop events from ITTNotify, it also
provides an event when a JITed function is compiled and placed into
memory. This handles those events in the ITTNotify collector. Sampled
addresses which are not resolved by BFD are checked against registered
JIT addresses. This allows resolution of sampled JITed Julia functions.


Former-commit-id: 325070c8572f448010f46a9a089758399603a25e